--- module1/src/main/java/net/broscorp/a_patform/README.md
@@ -1,6 +1,0 @@
-# Архитектура платформы java 
-
-При помощи языка [DOT](https://en.wikipedia.org/wiki/DOT_(graph_description_language)) нарисуй диаграмму связей между JVM, JRE, JIT, JDK.
-Покажи в каком порядке они включают друг друга, с чем работают, добавь краткое описание ответственности каждого компонента.
-
-Проверять результаты можно на сайте http://www.webgraphviz.com/. В репозиторий надо добавить файл .gv с разметкой.
