--- module1/src/main/java/net/broscorp/b_bytecode/README.md
@@ -1,17 +1,0 @@
-# Bytecode
-
-Это упражнение должно помочь тебе прочувствовать что такое байткод
-
-Создай в этой папке небольшой класс с main методом, который делает простые вещи
-* Инкремент переменной в цикле
-* Вызов метода
-* Печать приветствия в консоль
- 
-Напиши небольшой sh или cmd(если ты еще не достиг просветления) скрипт, который
-1. Компилирует класс
-2. Запускает .class файл, чтоб тот напечатал своё сообщение
-3. Дизассемблирует .class файл в байткод
-
-Посмотри на результат декомпиляции, попробуй менять код класса и посмотри как меняется байткод
-
-Добавь результаты декомпиляции разных варианатов в репозиторий.
