+++ module1/src/main/java/net/broscorp/boxing/README.md
@@ -1,0 +1,6 @@
+# Wrappers and boxing
+
+## Алгоритм выполнения
+
+1.	Написать тест, который покажет, что оберточные типы нельзя сравнивать через __==__.
+2.	Написать тест с использованием явного __boxing/unboxing__.
