+++ module1/src/main/java/net/broscorp/bytecode/README.md
@@ -1,0 +1,16 @@
+# Bytecode 
+_Это упражнение должно помочь тебе прочувствовать что такое __bytecode__._
+
+## Алгоритм выполнения
+
+1.	Создать в этой папке небольшой класс с __main__ методом, который делает простые вещи:
+    - инкремент переменной в цикле;
+    - вызов метода;
+    - печать приветствия в консоль.
+2.	Напиcать небольшой __sh__ или __cmd__ (если ты еще не достиг просветления) скрипт, который:
+    - компилирует класс;
+    - запускает __.class__ файл, чтоб тот напечатал своё сообщение;
+    - дизассемблирует __.class__ файл в __bytecode__.
+3.	Посмотреть на результат декомпиляции. Изменить код класса и _проанализировать_, как меняется __bytecode__.
+4.	Добавить результаты декомпиляции разных вариантов в репозиторий.
+
