+++ module1/src/main/java/net/broscorp/checkstyle/README.md
@@ -1,0 +1,13 @@
+# Checkstyle 
+_Сейчас мы разберемся с иснтструментом, позволяющим нам писать код в одном стиле._
+
+## Алгоритм выполнения
+
+1.	В папке `module1` выполнить `mvn clean install` билд упадет.
+2.	Заглянуть в `target/checkstyle-result.xml` 
+3.  Выполнить `mvn checkstyle:checkstyle` 
+3.	Посмотреть в консоли список ошибок.
+4.	Открыть файл `target/site/scheckstyle.html` это еще одно место, где можно увидеть все ошибки
+5.  Исправить все замечания плагина, билд должен начать проходить без проблем.
+ 
+Создавая пулл реквест для других заданий StyledClass можно удалить.
