+++ module1/src/main/java/net/broscorp/checkstyle/StyledClass.java
@@ -1,0 +1,7 @@
+package net.broscorp.checkstyle;
+
+public class StyledClass {
+    public static void MAIN(String[] args) {
+        System.out.println ("Hello checkstyle");
+    }
+}
