+++ module1/src/main/java/net/broscorp/classpath/README.md
@@ -1,0 +1,9 @@
+# Classpath
+
+## Алгоритм выполнения
+
+1.	Создать два класса.
+2.	В первом классе __main__ метод и чтение значения из консоли, во втором – вывод его на консоль.
+3.	С помощью инструментов __JDK__, без __maven__ и __среды разработки__ необходимо скомпилировать и упаковать __.class файлы__ в __.jar__ архив.
+4.	Добавить файл __.cmd__ или __.sh__ с командами это выполняющими.
+
