--- module1/src/main/java/net/broscorp/d_gc/README.md
@@ -1,10 +1,0 @@
-# Сборщик мусора
-Надо написать тест, который будет создавать большое количество объектов.
-Выводя сообщение в finalize и создавая/теряя ссылку на объект посмотреть когда срабатывает gc. 
-Глянуть JConsole. 
-
-Проверить что будет если при финализации снова сделать объект достижимым.
-
-Проверь будут ли удалены пары объектов, которые ссылаются друг на друга.
-
-Результаты экспериментов описать текстом в файле results.md, привести пример логов для иллюстрации.
