--- module1/src/main/java/net/broscorp/e_primitives/README.md
@@ -1,7 +1,0 @@
-# Примитивные типы данных
-Написать класс с тестами на junit.
-
-Добавить в него тесты, которые илюстрируют следующие вещи:
-* Переполнение численного типа
-* Конвертация между примитивными типами: long в int, float в double, показать какие могут возникать проблемы
-* Погрешность при работе с числами с плавающей точкой
