+++ module1/src/main/java/net/broscorp/equals/hashcode/README.md
@@ -1,0 +1,10 @@
+# equals and hashCode
+
+## Алгоритм выполнения
+
+1.	Прочитать __javadoc__ методов __equals__ и __hashCode__.
+2.	Сделать новый класс с несколькими полями, реализовать для него __toString__, __hashCode__ и __equals__.
+3.	Написать _Unit Test_, что проверяет инварианты __equals__ и __hashCode__ из __javadoc__.
+4.	Добавить _Unit Test_, перебором находящий объекты с совпадающими хешами. 
+5.	Напечатать примеры не равных объектов, чьи хеши совпадают. Вам надо напечатать "Вот объект А, вот объект В, они разные, но их хеш сопадает". И проверить, что их __equals__ даёт __false__, а __hashCode__ – возвращают одинаковые значения. Для выполнения задачи достаточно одной коллекции.
+
