+++ module1/src/main/java/net/broscorp/exceptions/README.md
@@ -1,0 +1,13 @@
+# Exceptions
+
+## Алгоритм выполнения
+
+1.	Написать класс теста, который воспроизводит следующие случаи:
+    - __try-with-resources__ который открывает файл, но получает __IOException__, превозмогает его и печатает прощание в __finally__.
+    - Пример кода, перехватывающего предков исключений.
+2.	При помощи __DOT__ нарисовать диаграмму __Throwable, Exception, Error, Runtime Exception__. Пометить какие из них декларируемые, какие нет.
+
+
+## Источники для выполнения задания:
+1.	https://en.wikipedia.org/wiki/DOT_(graph_description_language)
+2.	http://www.webgraphviz.com/ – проверка результатов.
