--- module1/src/main/java/net/broscorp/f_boxing/README.md
@@ -1,5 +1,0 @@
-# Wrappers and boxing
-
-Написать тест, который покажет что оберточные типы нельзя сравнивать через ==
-
-Написать тест с использованием явного боксинга/анбоксинга
