+++ module1/src/main/java/net/broscorp/gc/README.md
@@ -1,0 +1,10 @@
+# Garbage collection
+
+## Алгоритм выполнения
+
+1.	Написать тест, который будет создавать большое количество объектов. 
+2.	Выводя сообщение в __finalize__ и создавая/теряя ссылку на объект, посмотреть, когда срабатывает __gc__.
+3.	Просмотреть __JConsole__.
+4.	Проверить, что будет, если при финализации снова сделать объект достижимым.
+5.	Проверить будут ли удалены пары объектов, которые ссылаются друг на друга.
+6.	Результаты экспериментов описать текстом в файле __results.md__, привести пример логов для иллюстрации.
