+++ module1/src/main/java/net/broscorp/generics/README.md
@@ -1,0 +1,8 @@
+# Параметризированные типы/generics
+
+## Алгоритм выполнения
+
+1.	Добавить реализацию списка в класс __MyCoolList__, переделать его интерфейс на использование дженериков.
+2.	Наложить ограничение чтоб список можно было параметризировать только численными типами.
+3.	Покрыть код тестами.
+
