+++ module1/src/main/java/net/broscorp/git/README.md
@@ -1,0 +1,43 @@
+# Git commands
+_Задание поможет тебе "прощупать" Git_
+
+## Алгоритм выполнения
+
+1.	__Git rebase__
+      *	Создать 2 ветки (__master__ / __feature__)
+      *	Сделать 2-3 коммита в каждой из веток
+      *	Перенести ветку __feature__ на вершину ветки 
+      *	Логи сбросить в ответ на задание (__git reflog__)
+2.	__Склеить коммиты__
+      *	Создать ветку 
+      *	Сделать 2-3 коммита 
+      *	Склеить коммиты в один  
+      *	Логи сбросить в ответ на задание (__git reflog__)
+3.	__Git merge__
+      *	Создать ветку __master__
+      *	Сделать 2 коммита 
+      *	Создать новую ветку __feature__, переключиться на неё 
+      *	Сделать 2 коммита 
+      *	Выполнить merge __feature__ в __master__
+      *	Логи сбросить в ответ на задание (__git reflog__)
+4.	__Git revert__
+      *	Создать ветку
+      *	Сделать 3 коммита  
+      *	Выполнить __revert__ последнего коммита 
+      *	Логи сбросить в ответ на задание (__git reflog__)
+5.	__Git cherry-pick__
+      *	Создать 2 ветки (__master__ / __feature__)
+      *	Сделать 2-3 коммита в каждой из веток
+      *	Перенести коммит из __feature__ в __master__
+      *	Логи сбросить в ответ на задание (__git reflog__).
+
+
+## Источники для выполнения задания:
+1.	https://git-scm.com/book/en/v2
+2.	https://dzone.com/articles/top-20-git-commands-with-examples
+3.	https://learngitbranching.js.org/?locale=uk
+4.	https://www.atlassian.com/git/tutorials/rewriting-history/git-rebase
+5.	https://htmlacademy.ru/blog/boost/tools/how-to-squash-commits-and-why-it-is-needed
+6.	https://www.atlassian.com/git/tutorials/using-branches/git-merge
+7.	https://www.atlassian.com/git/tutorials/undoing-changes/git-revert
+8.	https://www.atlassian.com/git/tutorials/cherry-pick
