--- module1/src/main/java/net/broscorp/gvalueref/Foo.java
@@ -1,11 +1,0 @@
-package net.broscorp.gvalueref;
-
-import java.util.List;
-
-public class Foo {
-
-  public void foo(int i, List<String> s) {
-    s.add("d");
-    i += 15;
-  }
-}
