--- module1/src/main/java/net/broscorp/gvalueref/README.md
@@ -1,3 +1,0 @@
-# Передача по ссылке и значению
-
-Написать тесты на класс в этом пакете. В комментариях к тесту своими словами объяснить почему так происходит.
