--- module1/src/main/java/net/broscorp/hgenerics/README.md
@@ -1,7 +1,0 @@
-# Параметризированные типы/generics
-
-Добавить реализацию списка в класс MyCoolList, переделать его интерфейс на использование дженериков.
- 
-Наложить ограничение чтоб список можно было параметризировать только численными типами.
-
-Покрыть код тестами.
