--- module1/src/main/java/net/broscorp/i_equals_hashcode/README.md
@@ -1,9 +1,0 @@
-# equals and hashCode
-
-Прочитать javadoc методов equals и hashCode
-
-Сделать новый класс с несколькими полями, реализовать для него hashCode и equals.
-
-Написать тест, что проверяет инварианты equals и hashCode из javadoc.
- 
-Добавить тест, перебором находящий объекты(с реализованным toString) с совпадающими хешами. Напечатать примеры не равных объектов, чьи хеши совпадают. Вам надо напечатать "Вот объект А, вот объект В, они разные, но их хеш сопадает". И проверить что их equals даёт false а hashCode возвращают одинаковые значения. Для выполнения задачи достаточно одной коллекции.
