+++ module1/src/main/java/net/broscorp/inner/classes/README.md
@@ -1,0 +1,9 @@
+# Inner Classes
+
+## Алгоритм выполнения
+
+1.	Сделать коллекцию из прошлого задания __Iterable__.
+2.	Реализовать итератор через __InnerClass__.
+3.	Покрыть код тестами.
+4.	В _javadoc_ описать разницу между __статическим InnerClass__ и __обычным__.
+
