--- module1/src/main/java/net/broscorp/j_exceptions/README.md
@@ -1,7 +1,0 @@
-# Exceptions
-Написать класс теста, который воспроизводит следующие случаи:
-
-* try-with-resources который открывает файл, но получает IOException, превозмогает его и печатает прощание в finally 
-* Пример кода, перехватывающего предков исключений
-
-При помощи [DOT](https://en.wikipedia.org/wiki/DOT_(graph_description_language)) нарисовать диаграмму Throwable, Exception, Error, Runtime Exception. Пометить какие из них декларируемые, какие нет.
