--- module1/src/main/java/net/broscorp/k_classpath/README.md
@@ -1,9 +1,0 @@
-# Classpath
-
-Создать два класса
-
-В первом классе main метод и чтение значения из консоли, во втором вывод его на консоль.
-
-С помощью инструментов **JDK**, без maven и среды разработки необходимо скомпилировать и упаковать .class файлы в .jar архив.
-
-Добавить файл .cmd или .sh с командами это выполняющими.
