--- module1/src/main/java/net/broscorp/l_inner_classes/README.md
@@ -1,5 +1,0 @@
-# Inner Classes
-
-Сделать коллекцию из прошлого задания Iterable, реализовать итератор через иннер класс. Добавить тесты.
-
-В джавадоке к нему описать разницу между статическим иннер классом и обычным.
