--- module1/src/main/java/net/broscorp/m_override_overload/README.md
@@ -1,8 +1,0 @@
-# Overriding / Overloading
-
-1. Сделать интерфейс MathOperation с методом, что получает два аргумента double и возвращает результат операции - double.
-2. Добавить перегрузки этого метода, что работают c int а также с обертками, возвращается всегда double.
-3. Добавить реализации интерфеса для операций сложения и умножения.
-4. Покрыть оба класса тестами
-5. Добавить тест, что вызывает метод с одним аргументом примитивом а вторым - оберткой, посмотреть какая из реализаций вызовется. 
-6. В джавадоке класса написать своимим словами назначение аннотации `@Override`
