--- module1/src/main/java/net/broscorp/n_strings/README.md
@@ -1,23 +1,0 @@
-# String
-
-1. Написать метод, который определяет, является ли слово палиндромом.
-
-2. Написать методы, которые из словосочетания "Hello World" :
-    1. Оставляет только слово Hello
-    2. Удаляет все буквы "l" в строке
-    3. Меняет все буквы "о" на букву "q"
-    4. Выводит список слов и количество букв в соответствующем слове. 
-3. Написать метод, который получает два числа - количество багов и количество куплетов и возвращает текст песни
- 
-```
-%n little bugs in the code, 
-%n little bugs in the code. 
-Take one down, patch it around %m little bugs in the code.
-```
-Где m = n - 10 + random.nextInt(20) 
-
-Для следующего куплета обновляем n = m
-
-Чтоб результаты вызова метода были повторяемы - инициализировать random произведением параметров метода.
-
-4. Для всего кода сделать тесты.
