--- module1/src/main/java/net/broscorp/o_git/README.md
@@ -1,51 +1,0 @@
-# Git commands
-
-[Справочник по основным командам](https://git-scm.com/book/en/v2)
-
-[Примеры использования](https://dzone.com/articles/top-20-git-commands-with-examples)
-
-[Потыкать палочкой в git можно тут](https://learngitbranching.js.org/)
-
-## Задания
-### 1. Git rebase
-[Справка](https://www.atlassian.com/git/tutorials/rewriting-history/git-rebase)
-
-      • Создать 2 ветки (master / feature)
-      • Сделать 2-3 коммита в каждой из веток
-      • Перенести ветку feature  на вершину ветки 
-      • Логи сбросить в ответ на задание (git reflog)
-
-### 2. Склеить коммиты
-[Справка ](https://htmlacademy.ru/blog/useful/git/how-to-squash-commits-and-why-it-is-needed)
-
-      • Создать ветку 
-      • Сделать 2-3 коммита 
-      • Склеить коммиты в один  
-      • Логи сбросить в ответ на задание (git reflog)
-
-### 3. Git merge
-[Справка](https://www.atlassian.com/git/tutorials/using-branches/git-merge)
-
-      • Создать ветку master
-      • Сделать 2 коммита 
-      • Создать новую ветку feature, переключиться на неё 
-      • Сделать 2 коммита 
-      • Выполнить merge feature в  master
-      • Логи сбросить в ответ на задание (git reflog)
-          
-
-### 4. Git revert
-[Справка](https://www.atlassian.com/git/tutorials/undoing-changes/git-revert)
-
-      • Создать ветку
-      • Сделать 3 коммита  
-      • Выполнить revert последнего коммита 
-      • Логи сбросить в ответ на задание (git reflog)
-
-### 5. Git cherry-pick
-[Справка](https://www.atlassian.com/git/tutorials/cherry-pick)
-
-      • Создать 2 ветки (master / feature)
-      • Сделать 2-3 коммита в каждой из веток
-      • Перенести коммит из  feature в в master
-      • Логи сбросить в ответ на задание (git reflog)
