+++ module1/src/main/java/net/broscorp/override/overload/README.md
@@ -1,0 +1,11 @@
+# Overriding / Overloading
+
+## Алгоритм выполнения
+
+1.	Сделать интерфейс __MathOperation__ с методом, что получает два аргумента __double__ и возвращает результат операции – __double__.
+2.	Добавить перегрузки этого метода, что работают c __int__, а также с обертками, возвращается всегда __double__.
+3.	Добавить реализации интерфейса для операций сложения и умножения.
+4.	Покрыть оба класса тестами.
+5.	Добавить тест, что вызывает метод с одним аргументом примитивом, а вторым – оберткой. Посмотреть какая из реализаций вызовется.
+6.	В __javadoc__ класса написать своими словами назначение аннотации __@Override__.
+
