+++ module1/src/main/java/net/broscorp/patform/README.md
@@ -1,0 +1,15 @@
+# Архитектура платформы java 
+
+## Алгоритм выполнения
+
+1.	Нарисовать диаграмму связей между __JVM, JRE, JIT, JDK__ при помощи языка __DOT__.
+2.	Показать в каком порядке они включают друг друга, с чем работают. 
+3.	Добавить краткое описание ответственности каждого компонента.
+4.	В репозиторий добавить файл __.gv__ с разметкой.
+
+
+## Полезные источники для выполнения задания:
+1.	https://en.wikipedia.org/wiki/DOT_(graph_description_language)
+2.	http://www.webgraphviz.com/ – проверка результатов.
+
+
