+++ module1/src/main/java/net/broscorp/primitives/README.md
@@ -1,0 +1,8 @@
+# Примитивные типы данных
+
+## Алгоритм выполнения
+
+Написать класс с тестами на __Junit__. Добавить в него тесты, которые иллюстрируют следующие вещи:
+   *	переполнение численного типа;
+   *	конвертация между примитивными типами: __long__ в __int__, __float__ в __double__, показать какие могут возникать проблемы;
+   *	погрешность при работе с числами с плавающей точкой.
