+++ module1/src/main/java/net/broscorp/valueref/README.md
@@ -1,0 +1,6 @@
+# Передача по ссылке и значению
+
+## Алгоритм выполнения
+
+1.	Написать тесты на класс __Foo__. 
+2.	Написать комментарии к тесту – своими словами объяснить почему так происходит.
